pr3/DataBaseWrapperPy_Grupo09/requirements.txt
pyautogui~=0.9.54
pytesseract~=0.3.13
pillow~=11.0.0